--- packages/frontend-2/lib/presentations/composables/postSetup.ts
@@ -1,12 +1,0 @@
-import { useViewerUserActivityTracking } from '~/lib/viewer/composables/activity'
-
-const useActivityTrackingIntegration = () => {
-  useViewerUserActivityTracking({
-    trackInternallyOnly: true
-  })
-}
-
-export const usePresentationViewerPostSetup = () => {
-  if (import.meta.server) return
-  useActivityTrackingIntegration()
-}
